social_law_simulation/requirements.txt
highway-env>=1.8.0
gymnasium>=0.28.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml>=6.0